Add a ModelManager where select Model.. But it an sooo basic implementation..
